api/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
tensorflow==2.16.1
numpy==1.26.4
pillow==10.4.0
python-multipart==0.0.9
pydantic-settings==2.6.1
tf-keras==2.16.0
